New Calculator 4

+++ src/Calculator.java
@@ -3,16 +3,16 @@ public class Calculator {
 
     public static void main(String args[]) {
 
-        int x=0;
-        int y=0;
-        int op=0;
+        float x=0;
+        float y=0;
+        float op=0;
         char c='+';
         Scanner userInput= new Scanner(System.in);
         System.out.print("x = ");
-        x=userInput.nextInt();
+        x=userInput.nextFloat();
 
         System.out.print("y = ");
-        y=userInput.nextInt();
+        y=userInput.nextFloat();
         System.out.print("Selecteaza un operator = ");
         c=userInput.next().charAt(0);
